requirements.txt
# Testing Framework
pytest>=7.4.0
pytest-html>=3.2.0
pytest-json-report>=1.5.0
pytest-rerunfailures>=12.0  # 失败重试

# HTTP Requests
requests>=2.31.0

# Logging
loguru>=0.7.0  # 日志系统

# Environment Management
python-dotenv>=1.0.0  # 环境变量管理

# Database
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0  # ORM & 连接池
pymysql>=1.1.0  # MySQL 驱动（备用）

# Data Processing
pandas>=2.0.0
openpyxl>=3.1.0

# Excel Reading (if needed)
xlrd>=2.0.0

# Test Management Platform
streamlit>=1.31.0  # 测试管理平台Web框架
streamlit-authenticator>=0.3.3  # 平台登录认证
pyyaml>=6.0  # 解析 auth_config.yaml
bcrypt>=4.0.0  # 密码 hash
fpdf2>=2.8.0
playwright>=1.40.0
